tests refact
