+++ source/ConfigurationService.Persistence.Tests/OptionGroupsTests.cs
@@ -1,0 +1,6 @@
+﻿namespace ConfigurationService.Persistence.Tests
+{
+    public class OptionGroupsTests
+    {
+    }
+}
