--- source/ConfigurationService.Tests/DomainTests/ValueObjectsTests/ProjectNameTests.cs
@@ -1,39 +1,0 @@
-﻿using System;
-using ConfigurationService.Domain.ValueObjects;
-using ConfigurationService.Tests.TestSetup;
-using Xunit;
-
-namespace ConfigurationService.Tests.DomainTests.ValueObjectsTests
-{
-    public class ProjectNameTests
-    {
-        [Theory] 
-        [InlineData("")]
-        [InlineData(" ")]
-        [InlineData("12")]
-        [InlineData("3proj")]
-        [InlineData("_proj")]
-        [InlineData("-proj")]
-        [InlineData(null)]
-        [InlineData("na me")]
-        [InlineData(TestLiterals.Project.Name.Incorrect)]
-        public void New_IncorrectNames_Exception(string name)
-        {
-            Assert.Throws<ApplicationException>(() => new ProjectName(name));
-        }
-
-        [Theory]
-        [InlineData("proj")]
-        [InlineData("p")]
-        [InlineData("Proj")]
-        [InlineData("pRoj")]
-        [InlineData("pRoj3")]
-        [InlineData("pRoj_3--g")]
-        [InlineData("pRoj_3--g")]
-        [InlineData(TestLiterals.Project.Name.Correct)]
-        public void New_CorrectNames_Success(string name)
-        {
-            var n = new ProjectName(name);
-        }
-    }
-}
