+++ source/ConfigurationService.Tests/Fixtures/DbContextFixture.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using ConfigurationService.Persistence;
+
+namespace ConfigurationService.Tests.Fixtures
+{
+    public class DbContextFixture
+    {
+        public ConfigurationServiceContext Context { get; }
+
+        public DbContextFixture(Action<TestContextConfigurer<ConfigurationServiceContext>> builder = null)
+        {
+            var m = new MockedContext<ConfigurationServiceContext>(builder);
+            Context = m.Object;
+        }
+    }
+}
