--- source/ConfigurationService.Tests/TestSetup/Presets/NonRootOptionGroup.cs
@@ -1,28 +1,0 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using ConfigurationService.Tests.TestSetup.Stubs;
-
-namespace ConfigurationService.Tests.TestSetup.Presets
-{
-    public class NonRootOptionGroup : IEnumerable<object[]>
-    {
-        public IEnumerator<object[]> GetEnumerator()
-        {
-            var p = new TestableProject(Guid.NewGuid(), TestLiterals.Project.Name.Correct, TestLiterals.Project.ApiKeys.Correct);
-            var env = new TestableEnvironment(Guid.NewGuid(), TestLiterals.Environment.Name.Correct, p);
-            p.AddEnv(env);
-            var group = new TestableOptionGroup(Guid.NewGuid(), 
-                TestLiterals.OptionGroup.Name.Correct, 
-                TestLiterals.OptionGroup.Description.Correct, 
-                env, env.GetRootOptionGroop());
-
-            yield return new object[] { group };
-        }
-
-        IEnumerator IEnumerable.GetEnumerator()
-        {
-            return GetEnumerator();
-        }
-    }
-}
