--- source/ConfigurationService.Tests/TestSetup/Stubs/TestableProject.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using ConfigurationService.Domain.Entities;
-using ConfigurationService.Domain.ValueObjects;
-using Environment = ConfigurationService.Domain.Entities.Environment;
-
-namespace ConfigurationService.Tests.TestSetup.Stubs
-{
-    internal class TestableProject : Project
-    {
-        public TestableProject(Guid id, string name, Guid apiKey) : base(new ProjectName(name), new ApiKey(apiKey))
-        {
-            Id = id;
-        }
-
-        public void AddEnv(Environment env)
-        {
-            _environments.Add(env);
-        }
-    }
-}
